+++ src/main/java/soon/capstone/infrastructure/github/service/dto/response/GithubIssueCreateResponse.java
@@ -2,7 +2,7 @@ package soon.capstone.infrastructure.github.service.dto.response;
 
 public record GithubIssueCreateResponse(
 
-    int number
+    Long number
 
 ) {
 }
